feat(optimizer@m-10511126): cherry-pick optimizer related code from ce-dev

See merge request dp/ClickHouse!13295
# Conflicts:
#	.codebase/pipelines/ci.yaml
#	programs/server/Server.cpp
#	src/Common/LinkedHashMap.h
#	src/Core/Settings.h
#	src/Interpreters/Context.h
#	src/Interpreters/InterpreterDropStatsQuery.cpp
#	src/Interpreters/InterpreterSelectQueryUseOptimizer.cpp
#	src/Optimizer/EqualityInference.cpp
#	src/Optimizer/ImplementSetOperation.cpp
#	src/Optimizer/JoinGraph.cpp
#	src/Optimizer/PlanOptimizer.cpp
#	src/Optimizer/Property/PropertyDeriver.cpp
#	src/Optimizer/Property/PropertyMatcher.cpp
#	src/Optimizer/QueryUseOptimizerChecker.cpp
#	src/Optimizer/Rewriter/ColumnPruning.cpp
#	src/Optimizer/Rewriter/PredicatePushdown.cpp
#	src/Optimizer/Rule/Implementation/SetJoinDistribution.cpp
#	src/Optimizer/Rule/Rewrite/PushAggThroughJoinRules.cpp
#	src/Optimizer/Rule/Rewrite/RemoveRedundantRules.cpp
#	src/Optimizer/Rule/Transformation/InlineCTE.cpp
#	src/Optimizer/tests/explains/tpcds/q14.explain
#	src/Optimizer/tests/explains/tpcds/q75-2.explain
#	src/Optimizer/tests/explains/tpcds/q75.explain
#	src/Optimizer/tests/explains/tpcds_no_statistics/q4.explain
#	src/Optimizer/tests/explains/tpcds_no_statistics/q74.explain
#	src/Parsers/tests/gtest_stats_query.cpp
#	src/QueryPlan/AggregatingStep.cpp
#	src/QueryPlan/ApplyStep.cpp
#	src/QueryPlan/ExchangeStep.cpp
#	src/QueryPlan/JoinStep.h
#	src/QueryPlan/QueryPlan.cpp
#	src/QueryPlan/QueryPlanner.cpp
#	src/QueryPlan/TableScanStep.cpp
#	src/Statistics/CatalogAdaptorCnch.cpp
#	src/Statistics/StatsNdvBucketsExtendImpl.h
#	tests/queries/0_ce_problematic_stateless/00051_any_inner_join.sql
#	tests/queries/0_ce_problematic_stateless/00098_l_union_all.sql
#	tests/queries/0_ce_problematic_stateless/00316_rounding_functions_and_empty_block.sql
#	tests/queries/0_ce_problematic_stateless/00353_join_by_tuple.sql
#	tests/queries/0_ce_problematic_stateless/00441_nulls_in.sql
#	tests/queries/0_ce_problematic_stateless/00444_join_use_nulls.sql
#	tests/queries/0_ce_problematic_stateless/00445_join_nullable_keys.sql
#	tests/queries/0_ce_problematic_stateless/00597_push_down_predicate_long.sql
#	tests/queries/0_ce_problematic_stateless/01009_global_array_join_names.sql
#	tests/queries/0_ce_problematic_stateless/01029_early_constant_folding.sql
#	tests/queries/0_ce_problematic_stateless/01056_predicate_optimizer_bugs.sql
#	tests/queries/0_ce_problematic_stateless/01611_constant_folding_subqueries.reference
#	tests/queries/0_ce_problematic_stateless/01611_constant_folding_subqueries.sql
#	tests/queries/0_ce_problematic_stateless/40014_set_operation_implicit_conversion_2.sql
#	tests/queries/0_ce_problematic_stateless/40019_name_resolution_in_order_by.sql
#	tests/queries/0_stateless/00982_materialized_view_match_rewrite_optimizer.reference
#	tests/queries/0_stateless/00982_materialized_view_match_rewrite_optimizer.sql
#	tests/queries/0_stateless/02210_processors_profile_log_with_optimizer.reference
#	tests/queries/0_stateless/02210_processors_profile_log_with_optimizer.sql
#	tests/queries/0_stateless/02210_processors_profile_log_without_optimizer.reference
#	tests/queries/0_stateless/02210_processors_profile_log_without_optimizer.sql
#	tests/queries/0_stateless/40003_bucket_table_join.reference
#	tests/queries/0_stateless/40023_reorder_adjacent_windows.reference
#	tests/queries/0_stateless/41833_distributed_ddl_ansi.sql
#	tests/queries/0_stateless/45007_date_stats.reference
#	tests/queries/0_stateless/45007_date_stats.sql
#	tests/queries/0_stateless/45007_ip_stats.reference
#	tests/queries/0_stateless/45007_ip_stats.sql
#	tests/queries/4_cnch_stateless/40022_topn_filtering_opt.reference
#	tests/queries/4_cnch_stateless/40022_topn_filtering_opt.sql
#	tests/queries/4_cnch_stateless/40034_subcolumn_opt_through_union.reference
#	tests/queries/4_cnch_stateless/46012_optimize_multi_distinct.reference
#	tests/queries/4_cnch_stateless/46012_optimize_multi_distinct.sql
#	tests/queries/4_cnch_stateless/48015_simplify_multiIf_function.reference
#	tests/queries/4_cnch_stateless/48017_not_push_filter_into_any_join.reference
#	tests/queries/4_cnch_stateless_no_tenant/40022_topn_filtering_opt.reference
#	tests/queries/4_cnch_stateless_no_tenant/40034_subcolumn_opt_through_union.reference
#	tests/queries/4_cnch_stateless_no_tenant/46012_optimize_multi_distinct.reference
#	tests/queries/4_cnch_stateless_no_tenant/48015_simplify_multiIf_function.reference
#	tests/queries/4_cnch_stateless_no_tenant/48017_not_push_filter_into_any_join.reference

+++ src/Optimizer/tests/gtest_domain_translator.cpp
@@ -1213,7 +1213,6 @@ TEST_F(DomainTranslatorTest, testImplicitCastOnValueSide)
         context);
 }
 
-/*
 TEST_F(DomainTranslatorTest, testInPredicateWithInteger)
 {
     ContextMutablePtr context = Context::createCopy(getContext().context);
@@ -1229,7 +1228,6 @@ TEST_F(DomainTranslatorTest, testInPredicateWithBigint)
 
     testInPredicate(int64_sym, int64_type, field_1, field_2, context);
 }
- */
 /* TODO: can not support the expression which is 'symbol(narrow type) operator value(wide type)' unless use 'cast' function;
 TEST_F(DomainTranslatorTest, testInPredicateWithReal)
 {
@@ -1250,7 +1248,7 @@ TEST_F(DomainTranslatorTest, testInPredicateWithDouble)
 
     testInPredicateWithFloatingPoint(float64_sym, float64_type, Field(Float64(1.0)),  Field(Float64(2.0)),  Field(Float64(std::sqrt(-1.0))), context);
 }
-
+*/
 TEST_F(DomainTranslatorTest, testInPredicateWithShortDecimal)
 {
     ContextMutablePtr context = Context::createCopy(getContext().context);
@@ -1275,8 +1273,8 @@ TEST_F(DomainTranslatorTest, testInPredicateWithLongDecimal)
     testInPredicate(
         decimal_38_3_sym,
         decimal_38_3_type,
-        DecimalField<Decimal128>(10, 3),
-        DecimalField<Decimal128>(20, 3), context);
+        DecimalField<Decimal128>(Int128(10), 3),
+        DecimalField<Decimal128>(Int128(20), 3), context);
 }
 
 TEST_F(DomainTranslatorTest, testInPredicateWithVarchar)
@@ -1291,7 +1289,6 @@ TEST_F(DomainTranslatorTest, testInPredicateWithVarchar)
         "second",
         context);
 }
- */
 /*
 TEST_F(DomainTranslatorTest, testNonImplicitCastOnSymbolSide)
     public void testInPredicateWithEquitableType()
